Uses Template for future COurses
