+++ src/courses/templates/coursecontent/coursecontent.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import arrow from '../../../../assets/open.png';
+import arrow from '../../../assets/open.png';
 
-function CourseContent() {
+
+function CourseContent({ navs}) {
   const [open, setOpen] = React.useState(true);
   const onOpen = () => { 
     setOpen(prev => !prev)
@@ -16,6 +17,12 @@ function CourseContent() {
           flex text-gray-700
         `}>
         <Nav open={open} navs={navs} />
+        
+        {/* Course Content*/}
+        <div className={`w-full h-full bg-white flex flex-col justify-start items-center overflow-scroll px-2 pt-14`}>
+          <Header open={open} onOpen={() => onOpen()} />
+          <Course chap={102}/>
+        </div>
       </div>
     </>
   );
@@ -26,18 +33,51 @@ export default CourseContent;
 function Nav({ open, navs }) {
   return (
     <>
-      <ol
+      <div
         className={`
           ${ open ? '' : 'hidden' }
-          w-72 h-full bg-violet-200 
+          w-72 h-full bg-violet-900 
           pt-14 flex flex-col
         `}>
+        <div
+          className={`
+            text-white font-bold text-sm 
+            tracking-widest
+          `}>
+          <p>CONCEPTS</p>
+        </div>
         { 
           navs && navs.map( (nav, item) => (
-            <li key={item}><Link to={'test'} className='px-5 py-2 text-sm'>{section.name}</Link></li>
+            <Link to={'test'} key={item} className='px-5 py-2 text-sm text-white'>{nav}</Link>
           ))
         }
-      </ol>
+      </div>
     </>
+  );
+}
+
+function Header({ open, onOpen }) {
+  return (
+    <div className={`w-full flex justify-center items-center pt-1`}>
+      <div className={`w-10 text-left pl-2`} onClick={onOpen}><img src={arrow} alt='open' className={`${open ? 'rotate-180' : '' }`} /></div>
+      <div className={`w-full flex justify-center`}>
+        <div className={`w-full flex justify-end max-w-screen-md pr-5`}>
+          <p className={`text-sm uppercase tracking-widest font-semibold text-violet-500`}>Send Page Feedback</p>
+        </div>
+      </div>
+      <div className={`hidden md:block w-10 text-right pr-2`}>?</div>
+    </div>
+  )
+}
+
+function Course({ chap }) {
+  return (
+  <div className={`w-full max-w-screen-md flex flex-col gap-5`}>
+    <div className={`w-full border-b-2 border-violet-700 pt-3 text-2xl`}>Why React ?</div>
+    <div className={`w-full tracking-wide`}>
+      <p>From <a href='https://en.wikipedia.org/wiki/Function_composition_(computer_science)'>Wikipedia</a>, Composition is:</p>
+      <p className={`p-3 font-mono`}><em>to combine simple functions to build more complicated ones</em></p>
+    </div>
+  </div>
   );
 }
